Move dependency to mjolnir from valhalla_test to valhalla_gurka_test
