+++ bench/thor/routes.cc
@@ -85,7 +85,7 @@ constexpr float kMaxRange = 256;
 
 static void BM_UtrechtBidirectionalAstar(benchmark::State& state) {
   const auto config = build_config("generated-live-data.tar");
-  test::build_live_traffic_data(config);
+  gurka::build_live_traffic_data(config);
 
   std::mt19937 gen(0); // Seed with the same value for consistent benchmarking
   {
@@ -107,10 +107,10 @@ static void BM_UtrechtBidirectionalAstar(benchmark::State& state) {
       } else {
       }
     };
-    test::customize_live_traffic_data(config, generate_traffic);
+    gurka::customize_live_traffic_data(config, generate_traffic);
   }
 
-  auto clean_reader = test::make_clean_graphreader(config.get_child("mjolnir"));
+  auto clean_reader = gurka::make_clean_graphreader(config.get_child("mjolnir"));
 
   std::vector<valhalla::baldr::Location> locations;
 
@@ -183,7 +183,7 @@ static void BM_UtrechtBidirectionalAstar(benchmark::State& state) {
 void customize_traffic(const boost::property_tree::ptree& config,
                        baldr::GraphId& target_edge_id,
                        const int target_speed) {
-  test::build_live_traffic_data(config);
+  gurka::build_live_traffic_data(config);
   // Make some updates to the traffic .tar file.
   // Generate traffic data
   std::function<void(baldr::GraphReader&, baldr::TrafficTile&, int, baldr::TrafficSpeed*)>
@@ -198,7 +198,7 @@ void customize_traffic(const boost::property_tree::ptree& config,
       current->encoded_speed1 = target_speed >> 1;
     }
   };
-  test::customize_live_traffic_data(config, generate_traffic);
+  gurka::customize_live_traffic_data(config, generate_traffic);
 }
 
 BENCHMARK(BM_UtrechtBidirectionalAstar)->Unit(benchmark::kMillisecond);
@@ -242,7 +242,7 @@ void BM_GlobalFixedRandom(benchmark::State& state, const std::string& planet_pat
                         {{"additional_data", "mjolnir.traffic_extract", "mjolnir.tile_dir"}});
   config.put("mjolnir.tile_extract", planet_path);
 
-  auto clean_reader = test::make_clean_graphreader(config.get_child("mjolnir"));
+  auto clean_reader = gurka::make_clean_graphreader(config.get_child("mjolnir"));
 
   Options options;
   create_costing_options(options);
@@ -317,7 +317,7 @@ static void BM_GetSpeed(benchmark::State& state) {
   const auto tgt_speed = 50;
   customize_traffic(config, tgt_edge_id, tgt_speed);
 
-  auto clean_reader = test::make_clean_graphreader(config.get_child("mjolnir"));
+  auto clean_reader = gurka::make_clean_graphreader(config.get_child("mjolnir"));
 
   auto tile = clean_reader->GetGraphTile(baldr::GraphId(tgt_edge_id));
   if (tile == nullptr) {
@@ -349,7 +349,7 @@ static void BM_Sif_Allowed(benchmark::State& state) {
   auto tgt_speed = 100;
   customize_traffic(config, tgt_edge_id, tgt_speed);
 
-  auto clean_reader = test::make_clean_graphreader(config.get_child("mjolnir"));
+  auto clean_reader = gurka::make_clean_graphreader(config.get_child("mjolnir"));
 
   Options options;
   create_costing_options(options);
