+++ test/gurka/test_bidir_search.cc
@@ -22,10 +22,10 @@ TEST(StandAlone, exhaust_reverse_search) {
   const std::string tile_dir = "test/data/bidir_search";
   auto map = gurka::buildtiles(layout, ways, {}, {}, tile_dir);
   std::shared_ptr<baldr::GraphReader> reader =
-      test::make_clean_graphreader(map.config.get_child("mjolnir"));
+      gurka::make_clean_graphreader(map.config.get_child("mjolnir"));
 
   // mark AB, BC & CD not_thru in both directions
-  std::vector<GraphId> not_thru_edgeids;
+  std::vector<baldr::GraphId> not_thru_edgeids;
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "A", "B")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "B", "A")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "B", "C")));
@@ -33,7 +33,8 @@ TEST(StandAlone, exhaust_reverse_search) {
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "C", "D")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "D", "C")));
 
-  test::customize_edges(map.config, [&not_thru_edgeids](const GraphId& edgeid, DirectedEdge& edge) {
+  gurka::customize_edges(map.config, [&not_thru_edgeids](const baldr::GraphId& edgeid,
+                                                         baldr::DirectedEdge& edge) {
     if (std::find(not_thru_edgeids.begin(), not_thru_edgeids.end(), edgeid) !=
         not_thru_edgeids.end()) {
       edge.set_not_thru(true);
@@ -66,10 +67,10 @@ TEST(StandAlone, exhaust_forward_search) {
   const auto layout = gurka::detail::map_to_coordinates(ascii_map, 100);
   auto map = gurka::buildtiles(layout, ways, {}, {}, tile_dir);
   std::shared_ptr<baldr::GraphReader> reader =
-      test::make_clean_graphreader(map.config.get_child("mjolnir"));
+      gurka::make_clean_graphreader(map.config.get_child("mjolnir"));
 
   // mark CD, DE & EF not_thru in both directions
-  std::vector<GraphId> not_thru_edgeids;
+  std::vector<baldr::GraphId> not_thru_edgeids;
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "C", "D")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "D", "C")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "D", "E")));
@@ -77,7 +78,8 @@ TEST(StandAlone, exhaust_forward_search) {
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "E", "F")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "F", "E")));
 
-  test::customize_edges(map.config, [&not_thru_edgeids](const GraphId& edgeid, DirectedEdge& edge) {
+  gurka::customize_edges(map.config, [&not_thru_edgeids](const baldr::GraphId& edgeid,
+                                                         baldr::DirectedEdge& edge) {
     if (std::find(not_thru_edgeids.begin(), not_thru_edgeids.end(), edgeid) !=
         not_thru_edgeids.end()) {
       edge.set_not_thru(true);
@@ -109,10 +111,10 @@ TEST(StandAlone, failed_search) {
   const auto layout = gurka::detail::map_to_coordinates(ascii_map, 100);
   auto map = gurka::buildtiles(layout, ways, {}, {}, tile_dir);
   std::shared_ptr<baldr::GraphReader> reader =
-      test::make_clean_graphreader(map.config.get_child("mjolnir"));
+      gurka::make_clean_graphreader(map.config.get_child("mjolnir"));
 
   // mark CD, DE & EF not_thru in both directions
-  std::vector<GraphId> not_thru_edgeids;
+  std::vector<baldr::GraphId> not_thru_edgeids;
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "B", "C")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "C", "B")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "C", "D")));
@@ -120,7 +122,8 @@ TEST(StandAlone, failed_search) {
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "D", "E")));
   not_thru_edgeids.push_back(std::get<0>(gurka::findEdgeByNodes(*reader, layout, "E", "D")));
 
-  test::customize_edges(map.config, [&not_thru_edgeids](const GraphId& edgeid, DirectedEdge& edge) {
+  gurka::customize_edges(map.config, [&not_thru_edgeids](const baldr::GraphId& edgeid,
+                                                         baldr::DirectedEdge& edge) {
     if (std::find(not_thru_edgeids.begin(), not_thru_edgeids.end(), edgeid) !=
         not_thru_edgeids.end()) {
       edge.set_not_thru(true);
