+++ test/gurka/test_match.cc
@@ -210,18 +210,18 @@ protected:
     map.config.put("mjolnir.traffic_extract", "test/data/match_timedep/traffic.tar");
 
     // add live traffic
-    test::build_live_traffic_data(map.config);
-    test::customize_live_traffic_data(map.config, [&](baldr::GraphReader&, baldr::TrafficTile&, int,
+    gurka::build_live_traffic_data(map.config);
+    gurka::customize_live_traffic_data(map.config, [&](baldr::GraphReader&, baldr::TrafficTile&, int,
                                                       valhalla::baldr::TrafficSpeed* traffic_speed) {
       traffic_speed->overall_encoded_speed = 50 >> 1;
       traffic_speed->encoded_speed1 = 50 >> 1;
       traffic_speed->breakpoint1 = 255;
     });
 
-    test::customize_historical_traffic(map.config, [](DirectedEdge& e) {
+    gurka::customize_historical_traffic(map.config, [](baldr::DirectedEdge& e) {
       e.set_constrained_flow_speed(25);
       e.set_free_flow_speed(75);
-      std::array<float, kBucketsPerWeek> historical;
+      std::array<float, baldr::kBucketsPerWeek> historical;
       historical.fill(7);
       for (size_t i = 0; i < historical.size(); ++i) {
         // TODO: if we are in morning or evening set a different speed and add another test
@@ -232,7 +232,7 @@ protected:
     // tests below use saturday at 9am and thursday 4am (27 if for leap seconds)
     size_t second_of_week_constrained = 5 * 24 * 60 * 60 + 9 * 60 * 60 + 27;
     size_t second_of_week_freeflow = 3 * 24 * 60 * 60 + 4 * 60 * 60 + 27;
-    auto reader = test::make_clean_graphreader(map.config.get_child("mjolnir"));
+    auto reader = gurka::make_clean_graphreader(map.config.get_child("mjolnir"));
     for (auto tile_id : reader->GetTileSet()) {
       auto tile = reader->GetGraphTile(tile_id);
       for (const auto& e : tile->GetDirectedEdges()) {
