+++ test/gurka/test_use_living_streets.cc
@@ -43,10 +43,10 @@ protected:
         gurka::buildtiles(layout, ways, {}, {}, "test/data/gurka_use_living_streets");
 
     // set the same historical speeds for all roads
-    test::customize_historical_traffic(use_living_streets_map.config, [](baldr::DirectedEdge& e) {
+    gurka::customize_historical_traffic(use_living_streets_map.config, [](baldr::DirectedEdge& e) {
       e.set_free_flow_speed(40);
       e.set_constrained_flow_speed(40);
-      return std::array<float, kBucketsPerWeek>{};
+      return std::array<float, baldr::kBucketsPerWeek>{};
     });
   }
 };
