+++ test/test.h
@@ -5,7 +5,6 @@
 #include "baldr/rapidjson_utils.h"
 #include "baldr/traffictile.h"
 #include "config.h"
-#include "mjolnir/graphtilebuilder.h"
 
 #include <cmath>
 #include <fstream>
@@ -50,46 +49,5 @@ boost::property_tree::ptree
 make_config(const std::string& path_prefix,
             const std::unordered_map<std::string, std::string>& overrides = {},
             const std::unordered_set<std::string>& removes = {});
-
-/**
- * Generate a new GraphReader that doesn't re-use a previously
- * statically initizalized tile_extract member variable.
- *
- * Useful if you need to reload a tile extract within the same
- * process
- */
-std::shared_ptr<valhalla::baldr::GraphReader>
-make_clean_graphreader(const boost::property_tree::ptree& mjolnir_conf);
-
-/*************************************************************/
-// Creates an empty traffic file
-//
-// To actually customize the traffic data, use `customize_live_traffic_data`
-//
-/*************************************************************/
-void build_live_traffic_data(const boost::property_tree::ptree& config,
-                             uint32_t traffic_tile_version = valhalla::baldr::TRAFFIC_TILE_VERSION);
-
-/*************************************************************/
-// Helper function for customizing traffic data in unit-tests
-//
-// `setter_cb` is a callback that can modify traffic for each edge
-// when building traffic data
-/*************************************************************/
-using LiveTrafficCustomize = std::function<void(valhalla::baldr::GraphReader&,
-                                                valhalla::baldr::TrafficTile&,
-                                                int,
-                                                valhalla::baldr::TrafficSpeed*)>;
-
-void customize_live_traffic_data(const boost::property_tree::ptree& config,
-                                 const LiveTrafficCustomize& setter_cb);
-
-using HistoricalTrafficCustomize =
-    std::function<boost::optional<std::array<float, kBucketsPerWeek>>(DirectedEdge&)>;
-void customize_historical_traffic(const boost::property_tree::ptree& config,
-                                  const HistoricalTrafficCustomize& cb);
-
-using EdgesCustomize = std::function<void(const GraphId&, DirectedEdge&)>;
-void customize_edges(const boost::property_tree::ptree& config, const EdgesCustomize& setter_cb);
 
 } // namespace test
